move Test Template to testx/mocks/workbooksMock.ts
